+++ .github/scripts/telegram/parse_and_send_team_issues.py
@@ -473,78 +473,6 @@ def parse_team_issues(content):
     return teams
 
 
-def escape_markdown(text):
-    """
-    Escape special MarkdownV2 characters for Telegram, but preserve bold formatting and link structure.
-    
-    Args:
-        text (str): Text to escape
-        
-    Returns:
-        str: Escaped text
-    """
-    # For MarkdownV2, we need to escape these characters: _ * [ ] ( ) ~ ` > # + - = | { } . !
-    # But we want to preserve * for bold formatting, [ ] ( ) for links, and content inside backticks
-    
-    # First, protect inline code by temporarily replacing them
-    code_pattern = r'`([^`]+)`'
-    code_blocks = []
-    
-    def replace_code(match):
-        code_blocks.append(match.group(1))
-        return f"CODEPLACEHOLDER{len(code_blocks)-1}N"
-    
-    # Replace all inline code with placeholders
-    text = re.sub(code_pattern, replace_code, text)
-    
-    # Then protect links by temporarily replacing them
-    link_pattern = r'\[([^\]]+)\]\(([^)]+)\)'
-    links = []
-    
-    def replace_link(match):
-        links.append((match.group(1), match.group(2)))
-        return f"LINKPLACEHOLDER{len(links)-1}N"
-    
-    # Replace all links with placeholders
-    text = re.sub(link_pattern, replace_link, text)
-    
-    # Also protect standalone URLs (http/https) by temporarily replacing them
-    url_pattern = r'https?://[^\s\)]+'
-    urls = []
-    
-    def replace_url(match):
-        urls.append(match.group(0))
-        return f"URLPLACEHOLDER{len(urls)-1}N"
-    
-    # Replace all standalone URLs with placeholders
-    text = re.sub(url_pattern, replace_url, text)
-    
-    # Escape special characters for MarkdownV2 (single backslash)
-    # Characters that need escaping: _ * [ ] ( ) ~ ` > # + - = | { } . !
-    special_chars = ['_', '~', '`', '>', '#', '+', '-', '=', '|', '{', '}', '.', '!', '(', ')']
-    
-    for char in special_chars:
-        text = text.replace(char, f'\\{char}')
-    
-    # Restore standalone URLs (they should NOT be escaped)
-    for i, url in enumerate(urls):
-        text = text.replace(f"URLPLACEHOLDER{i}N", url)
-    
-    # Restore links, escaping special characters in link text but NOT in URLs
-    for i, (link_text, link_url) in enumerate(links):
-        # Escape special characters in link text only
-        escaped_text = link_text
-        for char in special_chars:
-            escaped_text = escaped_text.replace(char, f'\\{char}')
-        
-        # URLs should NOT be escaped (Telegram handles them correctly)
-        text = text.replace(f"LINKPLACEHOLDER{i}N", f"[{escaped_text}]({link_url})")
-    
-    # Restore inline code (content should NOT be escaped)
-    for i, code_content in enumerate(code_blocks):
-        text = text.replace(f"CODEPLACEHOLDER{i}N", f"`{code_content}`")
-    
-    return text
 
 
 def format_team_message(team_name, issues, team_responsible=None, muted_stats=None, show_diff=False):
@@ -615,9 +543,9 @@ def format_team_message(team_name, issues, team_responsible=None, muted_stats=No
         responsible = team_responsible[team_name]
         # Handle both single responsible and list of responsibles
         if isinstance(responsible, list):
-            responsible_str = " ".join(f"@{r.replace('_', '\\_')}" if not r.startswith('@') else r.replace('_', '\\_') for r in responsible)
+            responsible_str = " ".join(f"@{r}" if not r.startswith('@') else r for r in responsible)
         else:
-            responsible_str = f"@{responsible.replace('_', '\\_')}" if not responsible.startswith('@') else responsible.replace('_', '\\_')
+            responsible_str = f"@{responsible}" if not responsible.startswith('@') else responsible
         message += f"\n\nfyi: {responsible_str}"
     
     # Add empty line at the end for better readability
@@ -755,8 +683,7 @@ def send_team_messages(teams, bot_token, delay=2, max_retries=5, retry_delay=10,
         if not message.strip():
             continue
         
-        # Escape the entire message for MarkdownV2
-        message = escape_markdown(message)
+        # Message will be automatically escaped by send_telegram_message for MarkdownV2
         
         # Print final message before sending
         print(f"🔍 Final message for {team_name}:")
@@ -1118,13 +1045,12 @@ def send_period_updates(period, bot_token, team_channels, ydb_config, delay=2, m
             responsible = team_responsible[team_name]
             # Handle both single responsible and list of responsibles
             if isinstance(responsible, list):
-                responsible_str = " ".join(f"@{r.replace('_', '\\_')}" if not r.startswith('@') else r.replace('_', '\\_') for r in responsible)
+                responsible_str = " ".join(f"@{r}" if not r.startswith('@') else r for r in responsible)
             else:
                 responsible_str = f"@{responsible.replace('_', '\\_')}" if not responsible.startswith('@') else responsible.replace('_', '\\_')
             message += f"fyi: {responsible_str}\n\n"
         
-        # Escape the entire message for MarkdownV2
-        message = escape_markdown(message)
+        # Message will be automatically escaped by send_telegram_message for MarkdownV2
         
         # Print final message before sending
         print(f"🔍 Final {period}ly message for {team_name}:")
